+++ dev/refactor.py
@@ -1,7 +1,9 @@
 """Refactoring script to convert saved models to new names."""
+
 from __future__ import annotations
 
 import torch
+
 from matgl.models import MEGNet
 
 # model_path = "pretrained_models/MEGNet-MP-2019.4.1-BandGap-mfi"