+++ tests/conftest.py
@@ -8,17 +8,19 @@ Given that the fixtures are unlikely to be modified by the underlying code, the 
 "session". In the event that future tests are written that modifies the fixtures, these can be set to the default scope
 of "function".
 """
+
 from __future__ import annotations
 
-import matgl
 import pytest
 import torch
+from pymatgen.core import Lattice, Molecule, Structure
+from pymatgen.util.testing import PymatgenTest
+
+import matgl
 from matgl.ext.pymatgen import Molecule2Graph, Structure2Graph, get_element_list
 from matgl.graph.compute import (
     compute_pair_vector_and_distance,
 )
-from pymatgen.core import Lattice, Molecule, Structure
-from pymatgen.util.testing import PymatgenTest
 
 matgl.clear_cache(confirm=False)