+++ docs/changes.md
@@ -6,11 +6,21 @@ nav_order: 3
 
 # Change Log
 
+## 1.2.3
+- Fix dependency issues with DGL. Pinning to DGL<=2.1.0 for now, which have versions for all OSes.
+
+## 1.2.1
+- Bug fix for pbc dtype on Windows systems.
+
+## 1.2.0
+- Release of MatPES-based models.
+- Pin DGL and PyTorch dependencies to 2.2.0 to ensure compatibility with Mac.
+
 ## 1.1.3
 - Improve the memory efficiency and speed of three-body interactions. (@kenko911)
 - FrechetCellFilter is added for variable cell relaxation in Relaxer class. (@kenko911)
 - Smooth l1 loss function is added for training. (@kenko911)
-  
+
 ## 1.1.2
 - Move AtomRef Fitting to numpy to avoid bug (@BowenD-UCB)
 - NVE ensemble added (@kenko911)